pyproject.toml
[project]
name = "bitspeed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "enum34>=1.1.10",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]
